+++ paymentsheet-example/src/androidTest/java/com/stripe/android/lpm/TestPromptPay.kt
@@ -5,9 +5,6 @@ import com.stripe.android.BasePlaygroundTest
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.paymentsheet.example.playground.settings.Merchant
 import com.stripe.android.paymentsheet.example.playground.settings.MerchantSettingsDefinition
-import com.stripe.android.paymentsheet.example.playground.settings.Currency
-import com.stripe.android.paymentsheet.example.playground.settings.CurrencySettingsDefinition
-import com.stripe.android.paymentsheet.example.playground.settings.EnablePromptPaySettingsDefinition
 import com.stripe.android.paymentsheet.example.playground.settings.SupportedPaymentMethodsSettingsDefinition
 import com.stripe.android.test.core.AuthorizeAction
 import com.stripe.android.test.core.TestParameters
@@ -21,8 +18,6 @@ internal class TestPromptPay : BasePlaygroundTest() {
         authorizationAction = AuthorizeAction.ShowQrCodeThenPoll,
     ) { settings ->
         settings[MerchantSettingsDefinition] = Merchant.TH
-        settings[CurrencySettingsDefinition] = Currency.THB
-        settings[EnablePromptPaySettingsDefinition] = true
         settings[SupportedPaymentMethodsSettingsDefinition] = listOf(
             PaymentMethod.Type.Card,
             PaymentMethod.Type.PromptPay