+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/settings/PlaygroundSettings.kt
@@ -554,7 +554,6 @@ internal class PlaygroundSettings private constructor(
             TermsDisplaySettingsDefinition,
             PassiveCaptchaDefinition,
             AttestationOnIntentConfirmationDefinition,
-            EnablePromptPaySettingsDefinition,
             EnablePayPaySettingsDefinition,
             EnableTapToAddSettingsDefinition,
         )